Реализован UI и некоторые другие вещи

+++ UI/MainWindow.cs
@@ -23,13 +23,14 @@ namespace UI
         {
             InitializeComponent();
 
+            ulong appid = ulong.Parse(ConfigurationManager.AppSettings["AppIdForTest"]);
+
+#if DEBUG
             string username = ConfigurationManager.AppSettings["UsernameForTest"];
             string password = ConfigurationManager.AppSettings["PasswordForTest"];
-            ulong appid = ulong.Parse( ConfigurationManager.AppSettings["AppIdForTest"]);
-            string picFolder = @"C:\Users\admin\Desktop\Контент 1";
-
-            
-
+            textBoxLogin.Text = username;
+            textBoxPassword.Text = password;
+#endif
         }
 
         private void MainWindow_Load(object sender, EventArgs e)
@@ -53,29 +54,83 @@ namespace UI
             {
                 MessageBox.Show(ex.Message);
             }
+
+            
+
+            List<Group> groupLst = _vkHelper.GetGroupsWhereUserIsAdmin();
+
+            List<KeyValuePair<string, long>> groupsCBData = new List<KeyValuePair<string, long>>();
+            groupsCBData.Add(new KeyValuePair<string, long>("Не выбрано", -1));
+            foreach (var grp in groupLst)
+            {
+                groupsCBData.Add(new KeyValuePair<string, long>(grp.ToString(), grp.Id));
+            }
+
+            comboBoxGroups.DataSource = groupsCBData;
+            comboBoxGroups.DisplayMember = "Key";
+            comboBoxGroups.ValueMember = "Value";
+
+            dateTimePickerBeginDate.Format = DateTimePickerFormat.Custom;
+            dateTimePickerBeginDate.CustomFormat = "MM/dd/yyyy hh:mm:ss";
         }
 
         private void buttonLoad_Click(object sender, EventArgs e)
         {
             string contentFolder = textBoxContentPath.Text;
-
-            var picLst = FSClient.GetPicturesFromFolder(contentFolder);
+            if ( String.IsNullOrEmpty( contentFolder) ||  !Directory.Exists(contentFolder))
+            {
+                MessageBox.Show("Операция не может быть выполнена - не выбран каталог загрузки");
+                return;
+            }
+            
+            
+            long selectedGroupId = (long)comboBoxGroups.SelectedValue;
+            if (selectedGroupId == -1)
+            {
+                MessageBox.Show("Операция не может быть выполнена - не выбрана группа");
+                return;
+            }
 
             // алгоритм создания отложенных постов
-            long groupid = long.Parse(textBoxGroupId.Text);
-
+            var picLst = FSClient.GetPicturesFromFolder(contentFolder);
+            if (picLst.Count == 0)
+            {
+                MessageBox.Show("Операция не может быть выполнена - отсутствуют файлы для загрузки");
+                return;
+            }
+            
             // максимальное количество постов
-            int? maxPostCount = int.Parse( textBoxMaxPostCount.Text);
+            int maxPostCount = int.Parse( textBoxMaxPostCount.Text);
+            if (maxPostCount == 0)
+            {
+                MessageBox.Show("Операция не может быть выполнена - не установлено максимально возможное количетсво постов");
+                return;
+            }
+
+            // сколько постов на день
+            int maxPostOnDay = int.Parse(textBoxPostOnDayCount.Text);
+            if (maxPostOnDay == 0)
+            {
+                MessageBox.Show("Операция не может быть выполнена - не установлено максимально возможное количетсво постов в день");
+                return;
+            }
+
+            // дата начала
+            DateTime startDate = dateTimePickerBeginDate.Value;
+            if (startDate <= DateTime.Now)
+            {
+                MessageBox.Show("Операция не может быть выполнена - дата публикации должна быть минимум + 5 минут к текущему времени/дате");
+                return;
+            }
+
             // хэштэги
             string hashtags = textBoxPostHashtags.Text;
-            // сколько постов на день
-            int postOnDayCount = int.Parse( textBoxPostOnDayCount.Text);
+            
             // одна картинка - один пост
             // какой шаг между постами (часов)
             int postTimeGap = int.Parse( textBoxPostTimeStep.Text);
 
-            // дата начала
-            DateTime startDate = dateTimePickerBeginDate.Value;
+            
 
             // в каждом посте один анонимный опрос
             Poll poll = new Poll()
@@ -99,12 +154,13 @@ namespace UI
 
             foreach (var picPath in picLst)
             {
-                if (maxPostCount != null && postCounter == maxPostCount)
+                if ( postCounter == maxPostCount )
                 {
+                    MessageBox.Show("Операция выполнена");
                     break;
                 }
 
-                if (dailyPostCounter == postOnDayCount)
+                if (maxPostOnDay != -1 && dailyPostCounter == maxPostOnDay)
                 {
                     dailyPostCounter = 0;
                     dayCounter++;
@@ -114,14 +170,28 @@ namespace UI
                 // вычисление даты
                 DateTime postDate = dailyFirstPostDate.AddDays(dayCounter).AddHours(postTimeGap * dailyPostCounter);
 
-                _vkHelper.WallPost(groupid, postDate, hashtags, picPath, poll);
+                _vkHelper.WallPost(selectedGroupId, postDate, hashtags, picPath, poll);
 
-                File.Delete(picPath);
+                if (checkBoxDeleteFiles.Checked)
+                    File.Delete(picPath);
 
                 dailyPostCounter++;
                 postCounter++;
             }
 
+            if (maxPostCount == -1)
+            {
+                MessageBox.Show("Операция выполнена");
+            }
+        }
+
+        private void buttonSelectContentPath_Click(object sender, EventArgs e)
+        {
+            FolderBrowserDialog fbd = new FolderBrowserDialog();
+            if (fbd.ShowDialog() == DialogResult.OK)
+            {
+                textBoxContentPath.Text = fbd.SelectedPath;
+            }
         }
     }
 }
